+++ training/nvidia/dlrm-pytorch/config/config_common.py
@@ -1,5 +1,5 @@
 # DDP type: 'apex' or 'native'.
-ddp_type: str = "native"
+ddp_type: str = "apex"
 
 vendor = "nvidia"
 dist_backend = "nccl"
